bot/requirements.txt
numpy~=1.23.4
transformers~=4.19.4
requests~=2.28.1
lxml~=4.9.1
beautifulsoup4~=4.11.1
torch~=1.13.0
asyncio~=3.4.3
aioredis~=2.0.1
python-dotenv~=0.21.0
psycopg2~=2.9.5
PyYAML~=6.0
SQLAlchemy~=1.4.44
APScheduler~=3.9.1.post1
gensim~=3.6.0
redis~=4.3.4
textblob~=0.17.1
aiogram~=3.0.0b5